+++ src/main/java/com/portfolio/notification/domain/repository/NotificationRepository.java
@@ -2,8 +2,13 @@ package com.portfolio.notification.domain.repository;
 
 import com.portfolio.notification.domain.model.Notification;
 
+import java.util.Optional;
+import java.util.UUID;
+
 public interface NotificationRepository {
 
     Notification save(Notification notification);
+    Optional<Notification> findById(UUID id);
+
 
 }